+++ app/Domains/Identity/Controllers/UserController.php
@@ -7,6 +7,8 @@ use App\Domains\Identity\Requests\User\UpdateUserRequest;
 use App\Domains\Identity\Resources\UserResource;
 use App\Domains\Identity\Services\UserService;
 use App\Http\Controllers\Controller;
+use App\Models\Cliente;
+use App\Models\User;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
 
@@ -20,29 +22,35 @@ class UserController extends Controller
 
     public function index(): AnonymousResourceCollection
     {
+        $this->authorize('viewAny', User::class);
         return UserResource::collection($this->userService->findAll());
     }
 
-    public function show(int $id): UserResource
+    public function show(User $user): UserResource
     {
-        $user = $this->userService->find($id);
+        $this->authorize('view', $user);
+        $this->authorize('view', User::class);
+        $user = $this->userService->find($user->id);
         return new UserResource($user);
     }
 
     public function store(StoreUserRequest $request): UserResource
     {
+        $this->authorize('create', User::class);
         $user = $this->userService->create($request->validated());
         return new UserResource($user);
     }
 
-    public function update(int $id, UpdateUserRequest $request): UserResource
+    public function update(User $user, UpdateUserRequest $request): UserResource
     {
-        $user = $this->userService->update( $request->validated(), $id);
+        $this->authorize('update', $user);
+        $user = $this->userService->update( $request->validated(), $user->id);
         return new UserResource($user);
     }
 
     public function destroy(int $id): JsonResponse
     {
+        $this->authorize('delete', User::class);
         $this->userService->delete($id);
         return response()->json(['message' => 'Usuário deletado com sucesso!']);
     }
